Imposed strict width typing on the sample C application.

+++ src/timer.h
@@ -3,6 +3,7 @@
 
 #include <chrono>
 #include <ratio>
+#include "consts.h"
 
 namespace llce {
 
@@ -11,15 +12,15 @@ class timer {
 
     /// Constructors ///
 
-    timer( size_t pFPS = 60 );
+    timer( uint32_t pFPS = 60 );
 
     /// Class Functions ///
 
     void split();
     void wait();
 
-    double fps( size_t pNumFrames = 1 ) const;
-    double dt( size_t pNumFrames = 1 ) const;
+    float64_t fps( uint32_t pNumFrames = 1 ) const;
+    float64_t dt( uint32_t pNumFrames = 1 ) const;
 
     private:
 
@@ -28,7 +29,7 @@ class timer {
     using Clock = std::chrono::high_resolution_clock;
     using ClockPoint = decltype( Clock::now() );
     using ClockDuration = decltype( Clock::now() - Clock::now() );
-    using SecDuration = std::chrono::duration<long double, std::ratio<1>>;
+    using SecDuration = std::chrono::duration<float64_t, std::ratio<1>>;
 
     /// Class Fields ///
